+++ src/Ema/Server.hs
@@ -239,6 +239,8 @@ wsClientShim =
         function init(reconnecting) {
           console.log("ema: Opening ws conn");
           window.connecting();
+          // Awaiting initial HTML for current route
+          var awaitingInitialHtml = !reconnecting;
           var ws = new WebSocket("ws://" + window.location.host);
 
           // Call this, then the server will send update *once*. Call again for
@@ -258,6 +260,7 @@ wsClientShim =
               if (origin) {
                 if (window.location.host === origin.host) {
                   window.history.pushState({}, "", origin.pathname);
+                  awaitingInitialHtml = true;
                   switchRoute(origin.pathname);
                   e.preventDefault();
                 };
@@ -297,7 +300,13 @@ wsClientShim =
             } else {
               setHtml(document.documentElement, evt.data);
               reloadScripts(document.documentElement);
-              window.scrollTo({ top: 0});
+              console.log(awaitingInitialHtml);
+              if (awaitingInitialHtml) {
+                // This is a new route switch; scroll up.
+                window.scrollTo({ top: 0});
+              } else {
+                awaitingInitialHtml = false;
+              }
               watchCurrentRoute();
             };
           };