+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Avoid unncessary DOM patch on page load
   - Handle invalid routes gracefully without breaking websocket
   - Re-add `<script>` tags on hot reload
+  - Scroll to top on route switches
 - `Asset` type
   - Introduce the `Asset` type to distinguishing between static files and generated files. The later can be one of `Html` or `Other`, allowing the live server to handle them sensibly.
   - `Ema` typeclass: